Refactor: pass signout toogle state and the handleSignout function as props
